Fixed lots of compiler warnings.  Everything commented out in this commit should be safe to erase.
git-svn-id: svn+ssh://svn.code.sf.net/p/snowstorm/code/openglad/trunk@1591 636a839e-d62e-0410-834d-a1d0aacdb804

+++ src/parser.cpp
@@ -45,7 +45,7 @@ bool cfg_store::parse(const char *filename)
 	}
 	while(std::getline(in, line, '\n'))
 	{
-		int pos = line.find(';');
+		size_t pos = line.find(';');
 		if (pos != string::npos)
 			line.erase(pos);
 		if((pos=line.find('[')) != std::string::npos)
